fix bad links in readme
readme links were missing or just plain wrong in some cases. Accidentally had someone print me the wrong piece as a result!

Updated to point to the new links and added the studio display. Didn't link to the XDR yet as it said in the commit it was tentative so I left it off the readme for now.